Update face extraction script

+++ scripts/extract_facenet.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
+import argparse
 import contextlib
 import json
 import os
-import sys
 import time
-from multiprocessing.pool import ThreadPool, Pool
 from multiprocessing import Process
+from multiprocessing.pool import Pool, ThreadPool
 
 import torch
 import torch.backends.cudnn as cudnn
@@ -19,20 +19,27 @@ import pretorched
 from data import VideoFolder, video_collate_fn
 from models import FaceModel, deepmmag
 from pretorched.runners.utils import AverageMeter, ProgressMeter
+from pretorched.utils import str2bool
 
-try:
-    PART = sys.argv[1]
-except IndexError:
-    PART = 'dfdc_train_part_3'
 VIDEO_ROOT = os.path.join(os.environ['DATA_ROOT'], 'DeepfakeDetection', 'videos')
 FACE_ROOT = os.path.join(os.environ['DATA_ROOT'], 'DeepfakeDetection', 'facenet_frames')
-VIDEO_DIR = os.path.join(VIDEO_ROOT, PART)
-FACE_DIR = os.path.join(FACE_ROOT, PART)
+
+
+def parse_args():
+    parser = argparse.ArgumentParser(description='Face Extraction')
+    parser.add_argument('--part', type=str, default='dfdc_train_part_0')
+    parser.add_argument('--magnify_motion', default=False, type=str2bool)
+    parser.add_argument('--overwrite', default=False, type=str2bool)
+    parser.add_argument('--remove_frames', default=True, type=str2bool)
+    args = parser.parse_args()
+    args.video_dir = os.path.join(VIDEO_ROOT, args.part)
+    args.face_dir = os.path.join(FACE_ROOT, args.part)
+    return args
 
 
 def main(video_dir, face_dir, size=360, margin=100, fdir_tmpl='face_{}', tmpl='{:06d}.jpg',
          metadata_fname='face_metadata.json', step=1, batch_size=1, chunk_size=300, num_workers=2,
-         overwrite=False, remove_frames=True, magnify_motion=False):
+         overwrite=False, remove_frames=True, magnify_motion=False, **kwargs):
 
     os.makedirs(face_dir, exist_ok=True)
     device = 'cuda' if torch.cuda.is_available() else 'cpu'
@@ -124,7 +131,8 @@ def save_face_data(save_dir, face_images, run_motion_mag=None, size=360, margin=
             mm_out_dir = run_motion_mag(video=video_path, output=video_path + '_mm')
         if remove_frames:
             os.system(f'rm -rf {video_path}')
-            os.system(f'rm -rf {mm_out_dir}')
+            if run_motion_mag is not None:
+                os.system(f'rm -rf {mm_out_dir}')
 
 
 def frames_to_video(video_path):
@@ -186,4 +194,5 @@ def filter_filenames(video_filenames, face_dir):
 
 
 if __name__ == '__main__':
-    main(VIDEO_DIR, FACE_DIR)
+    args = parse_args()
+    main(**vars(args))
